timer fix

+++ Geometric2/Global/GlobalPhysicsData.cs
@@ -21,8 +21,8 @@ namespace Geometric2.Global
         public InitialConditionsData InitialConditionsData = new InitialConditionsData();
 
         //Visualization Data
-        public float alfaAngleInRadian = 0.5f;
-        public float diagonalRoundInRadian = 0.0f;
-        public float yRoundInRadian = 0.0f;
+        public decimal alfaAngleInRadian = 0.5M;
+        public decimal diagonalRoundInRadian = 0.0M;
+        public decimal yRoundInRadian = 0.0M;
     }
 }
